+++ src/render/wireframe_cache.js
@@ -31,8 +31,8 @@ export class WireframeDebugCache {
 
         const bufSize = gl.getBufferParameter(gl.ELEMENT_ARRAY_BUFFER, gl.BUFFER_SIZE);
         const bufTmp = new ArrayBuffer(bufSize);
-        const intView = new Int16Array(bufTmp);
-        gl.getBufferSubData(gl.ELEMENT_ARRAY_BUFFER, 0, new Int16Array(bufTmp));
+        const intView = new Int32Array(bufTmp);
+        gl.getBufferSubData(gl.ELEMENT_ARRAY_BUFFER, 0, new Int32Array(bufTmp));
 
         const lineIndexArray = new LineIndexArray();
